hotfix

+++ src/server/scripts/Northrend/Ulduar/Ulduar/instance_ulduar.cpp
@@ -1052,7 +1052,7 @@ public:
             }
             else if (unit->GetTypeId() == TYPEID_UNIT && unit->GetAreaId() == 4656 /*Conservatory of Life*/)
             {
-                if (time(nullptr) > (static_cast<uint64>(m_conspeedatoryAttempt) + DAY))
+                if (time(nullptr) > (m_conspeedatoryAttempt + DAY))
                 {
                     DoStartTimedAchievement(ACHIEVEMENT_TIMED_TYPE_EVENT, 21597 /*CON-SPEED-ATORY_TIMED_CRITERIA*/);
                     m_conspeedatoryAttempt = time(nullptr);
